# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This API provides endpoints to manage films and register users.\n[Read me](https://github.com/migalpha/kentech-films)",
        "title": "Kentech-Films",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/csv/films": {
            "get": {
                "description": "Export all films data to csv file.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Export all films data to csv file.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.emptyResponse"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Import films data from csv file.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Import films data from csv file.",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/films": {
            "get": {
                "description": "Allow to get all films by some filters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Get all films.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "film title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "film genre",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "film release year",
                        "name": "release_year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns films",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/film.Film"
                            }
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Allow to register a new film .",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Create a new film.",
                "parameters": [
                    {
                        "description": "create film",
                        "name": "film",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.createFilmRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http.createFilmResponse"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/films/{film_id}": {
            "get": {
                "description": "Given a film_id returns this film.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Get a specific film.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Film id",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns film",
                        "schema": {
                            "$ref": "#/definitions/film.Film"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Allow to remove a film from records.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Destroy a film from records.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Destroy film",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "error 403",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Allow to update one or many fields of films.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Films"
                ],
                "summary": "Allow to update one or many fields of films.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "film id",
                        "name": "film_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data to update",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateFilmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "error 400",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    },
                    "403": {
                        "description": "error 403",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error 500",
                        "schema": {
                            "$ref": "#/definitions/http.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "film.Film": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "integer"
                },
                "director": {
                    "type": "string",
                    "example": "Quentin Tarantino"
                },
                "genre": {
                    "type": "string",
                    "example": "action"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "release_year": {
                    "type": "integer",
                    "example": 2009
                },
                "starring": {
                    "type": "string",
                    "example": "Brad Pitt, Christoph Waltz, Michael Fassbender"
                },
                "sypnosis": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Inglourious Basterds"
                }
            }
        },
...
        "http.createFilmRequest": {
            "type": "object",
            "required": [
                "director",
                "genre",
                "release_year",
                "starring",
                "title"
            ],
            "properties": {
                "created_by": {
                    "type": "integer"
                },
                "director": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Quentin Tarantino"
                },
                "genre": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "action"
                },
                "release_year": {
                    "type": "integer",
                    "example": 2009
                },
                "starring": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Brad Pitt, Christoph Waltz, Michael Fassbender"
                },
                "sypnosis": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Inglourious Basterds"
                }
            }
        },
        "http.createFilmResponse": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "integer",
                    "example": 1
                },
                "director": {
                    "type": "string",
                    "example": "Quentin Tarantino"
                },
                "genre": {
                    "type": "string",
                    "example": "action"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "release_year": {
                    "type": "integer",
                    "example": 2009
                },
                "starring": {
                    "type": "string",
                    "example": "Brad Pitt, Christoph Waltz, Michael Fassbender"
                },
                "sypnosis": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Inglourious Basterds"
                }
            }
        },
        "http.emptyResponse": {
            "type": "object"
        },
        "http.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "http.updateFilmRequest": {
            "type": "object",
            "properties": {
                "director": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Quentin Tarantino"
                },
                "genre": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "action, comedy, war"
                },
                "release_year": {
                    "type": "integer",
                    "example": 2009
                },
                "starring": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Brad Pitt, Christoph Waltz, Michael Fassbender"
                },
                "sypnosis": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Inglourious Basterds"
                }
            }
        }
    }
}